api/requirements.txt
django
djangorestframework
django-filter
markdown
#rest-pandas
psycopg2
